Extract CmdBlockHelper to avoid compilations failures on i386.
